// ui/src/api/model/transferRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Bank System Core API REST
 * Documentación de la API REST del Core Bancario
 * OpenAPI spec version: 0.0.1
 */

export interface TransferRequest {
  /** @minimum 0 */
  amount: number;
  targetBankAccountId: string;
  /** @minLength 1 */
  concept: string;
  saveAsTrusted?: boolean;
}